+++ CatEye.Core/StageOperations/LimitSize/LimitSizeStageOperation.cs
@@ -20,7 +20,7 @@ namespace CatEye.Core
 		
 		public override double CalculateEfforts (IBitmapCore hdp)
 		{
-			return hdp.Width * hdp.Height;
+			return (double)hdp.Width * hdp.Height;
 		}
 		
 		public override void OnDo (IBitmapCore hdp)
